// src/modules/NotepadModule/redux/actions/Enum.ts
export enum ConstantsNotepad {
    GET_NOTEPADS = "GET_NOTEPADS",
    GET_NOTEPADS_SUCCESS = "GET_NOTEPADS_SUCCESS",
    GET_NOTEPADS_FAILED = "GET_NOTEPADS_FAILED",
    REMOVE_NOTEPAD_BY_ID = "REMOVE_NOTEPAD_BY_ID",
    REMOVE_NOTEPAD_BY_ID_SUCCESS = "REMOVE_NOTEPAD_BY_ID_SUCCESS",
    ADD_NOTEPAD = "ADD_NOTEPAD",
    ADD_NOTEPAD_SUCCESS = "ADD_NOTEPAD_SUCCESS",
    EDIT_NOTEPAD = "EDIT_NOTEPAD",
    EDIT_NOTEPAD_SUCCESS = "EDIT_NOTEPAD_SUCCESS",
    GET_NOTEPAD_BY_ID = "GET_NOTEPAD_BY_ID",
    GET_NOTEPAD_BY_ID_SUCCESS = "GET_NOTEPAD_BY_ID_SUCCESS",
    REMOVE_NOTE_BY_ID = "REMOVE_NOTE_BY_ID",
    REMOVE_NOTE_BY_ID_SUCCESS = "REMOVE_NOTE_BY_ID_SUCCESS",
}